MapRoulette: migrate new PR specs from JS to TypeScript
- Convert util/ui/service MapRoulette Vitest specs to .ts with strict typing
Ping https://github.com/tordans/iD/pull/4
