+++ test/spec/util/maproulette_pin_loc.ts
@@ -1,0 +1,109 @@
+import {
+  mapRouletteLineFeatures,
+  snapMapRoulettePinLoc,
+} from '../../../modules/util/maproulette_pin_loc';
+
+describe('maproulette_pin_loc', () => {
+  let curvedLine: any;
+
+  beforeEach(() => {
+    // C-shaped line: geometric center of the endpoints/bbox sits off the line.
+    curvedLine = {
+      type: 'FeatureCollection',
+      features: [{
+        type: 'Feature',
+        properties: {},
+        geometry: {
+          type: 'LineString',
+          coordinates: [
+            [0, 0],
+            [0, 2],
+            [2, 2],
+            [2, 0],
+          ],
+        },
+      }],
+    };
+  });
+
+  describe('mapRouletteLineFeatures', () => {
+    it('returns only LineString and MultiLineString features', () => {
+      const mixed = {
+        type: 'FeatureCollection',
+        features: [
+          { type: 'Feature', geometry: { type: 'Point', coordinates: [1, 1] }, properties: {} },
+          { type: 'Feature', geometry: { type: 'LineString', coordinates: [[0, 0], [1, 1]] }, properties: {} },
+          { type: 'Feature', geometry: { type: 'MultiLineString', coordinates: [[[0, 0], [1, 0]]] }, properties: {} },
+        ],
+      };
+      const lines = mapRouletteLineFeatures(mixed);
+      expect(lines).toHaveLength(2);
+      expect(lines[0].geometry.type).toBe('LineString');
+      expect(lines[1].geometry.type).toBe('MultiLineString');
+    });
+
+    it('returns empty when there are no line features', () => {
+      expect(mapRouletteLineFeatures(null)).toEqual([]);
+      expect(mapRouletteLineFeatures({
+        type: 'FeatureCollection',
+        features: [{
+          type: 'Feature',
+          geometry: { type: 'Point', coordinates: [1, 1] },
+          properties: {},
+        }],
+      })).toEqual([]);
+    });
+  });
+
+  describe('snapMapRoulettePinLoc', () => {
+    it('returns the original loc when there is no line geometry', () => {
+      expect(snapMapRoulettePinLoc([1, 1], null)).toEqual([1, 1]);
+      expect(snapMapRoulettePinLoc([1, 1], {
+        type: 'FeatureCollection',
+        features: [{
+          type: 'Feature',
+          geometry: { type: 'Point', coordinates: [9, 9] },
+          properties: {},
+        }],
+      })).toEqual([1, 1]);
+    });
+
+    it('snaps an off-line centerpoint onto the LineString', () => {
+      // Center of the C sits near (1,1), off the vertical/horizontal segments.
+      const snapped = snapMapRoulettePinLoc([1, 1], curvedLine);
+      // turf may snap to top or side — must lie on the polyline.
+      const onLeft = Math.abs(snapped[0] - 0) < 1e-6 && snapped[1] >= -1e-6 && snapped[1] <= 2 + 1e-6;
+      const onTop = Math.abs(snapped[1] - 2) < 1e-6 && snapped[0] >= -1e-6 && snapped[0] <= 2 + 1e-6;
+      const onRight = Math.abs(snapped[0] - 2) < 1e-6 && snapped[1] >= -1e-6 && snapped[1] <= 2 + 1e-6;
+      expect(onLeft || onTop || onRight).toBe(true);
+      // And not the off-line center.
+      expect(snapped[0] === 1 && snapped[1] === 1).toBe(false);
+    });
+
+    it('ignores Point features when snapping mixed collections', () => {
+      const mixed = {
+        type: 'FeatureCollection',
+        features: [
+          {
+            type: 'Feature',
+            geometry: { type: 'Point', coordinates: [5, 5] },
+            properties: {},
+          },
+          curvedLine.features[0],
+        ],
+      };
+      const snapped = snapMapRoulettePinLoc([1, 1], mixed);
+      expect(snapped[0] === 5 && snapped[1] === 5).toBe(false);
+      const onLeft = Math.abs(snapped[0] - 0) < 1e-6;
+      const onTop = Math.abs(snapped[1] - 2) < 1e-6;
+      const onRight = Math.abs(snapped[0] - 2) < 1e-6;
+      expect(onLeft || onTop || onRight).toBe(true);
+    });
+
+    it('leaves a point already on the line unchanged (within snap)', () => {
+      const snapped = snapMapRoulettePinLoc([0, 1], curvedLine);
+      expect(snapped[0]).toBeCloseTo(0, 5);
+      expect(snapped[1]).toBeCloseTo(1, 5);
+    });
+  });
+});
